API: Rename PyData backend to Numba
