+++ sparse/numba_backend/tests/conftest.py
@@ -3,6 +3,6 @@ import pytest
 
 @pytest.fixture(scope="session")
 def rng():
-    from sparse.pydata_backend._utils import default_rng
+    from sparse.numba_backend._utils import default_rng
 
     return default_rng
