+++ sparse/tests/conftest.py
@@ -3,7 +3,7 @@ import sparse
 import pytest
 
 
-@pytest.fixture(scope="session", params=[sparse.BackendType.PyData, sparse.BackendType.Finch])
+@pytest.fixture(scope="session", params=[sparse.BackendType.Numba, sparse.BackendType.Finch])
 def backend(request):
     with sparse.Backend(backend=request.param):
         yield request.param
